Player representation and game flow changed. (Issue #2)
All changes lead to possibility of surrending game, because this result
of game is known only at the moment of surrending). It is also necessary
to know if player won because of score or because of blackjack. All
those circumstances influence mechanism of rewarding or bet loosing.

Because things are becomming more complex there was also some additional
documentation added into the "doc" folder.